add confirmation modal

+++ src/app/app.component.ts
@@ -1,10 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { HeaderComponent } from './components/header/header.component';
 import { MatDialog } from '@angular/material/dialog';
-import {
-  User,
-  UserAddEditComponent,
-} from './components/user-add-edit/user-add-edit.component';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { UsersService } from './service/users.service';
 import { ViewChild } from '@angular/core';
@@ -16,6 +12,11 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { DatePipe } from '@angular/common';
 import { MatIconModule } from '@angular/material/icon';
 import { CommonModule } from '@angular/common';
+import { ConfirmationModalComponent } from './components/confirmation-modal/confirmation-modal.component';
+import {
+  User,
+  UserAddEditComponent,
+} from './user-add-edit/user-add-edit.component';
 
 @Component({
   selector: 'app-root',
@@ -35,6 +36,9 @@ import { CommonModule } from '@angular/common';
   styleUrl: './app.component.scss',
 })
 export class AppComponent implements OnInit {
+  title(title: any) {
+    throw new Error('Method not implemented.');
+  }
   displayedColumns: string[] = [
     'firstName',
     'lastName',
@@ -78,17 +82,32 @@ export class AppComponent implements OnInit {
     });
   }
 
-  deleteUser(id: number) {
-    this.userService.deleteUser(id).subscribe({
-      next: (res) => {
-        this.getUsersList();
-      },
+  deleteUser(id: number, userName: string): void {
+    const dialogRef = this.dialog.open(ConfirmationModalComponent, {
+      data: { userName: userName },
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.userService.deleteUser(id).subscribe({
+          next: () => {
+            this.getUsersList();
+          },
+        });
+      }
     });
   }
 
   openEditForm(data: User) {
-    this.dialog.open(UserAddEditComponent, {
+    const dialogRef = this.dialog.open(UserAddEditComponent, {
       data,
+    });
+    dialogRef.afterClosed().subscribe({
+      next: (val) => {
+        if (val) {
+          this.getUsersList();
+        }
+      },
     });
   }
 }
